refact(step2): rename utils to constants
